Move unit tests to integration because vscode deps
